Added ability to modify button font size directly
Students have had trouble in the past changing the font size on their DPEAButtons. This is because the font size belonged to a label overlaid on the DPEAButton and accessing this label was difficult. Now the label's font_size is directly tied to the font_size attribute of the DPEAButton that students indicate in their kv files.